requirements.txt
torch>=1.12
torchvision
numpy
Pillow
opencv-python
lpips
streamlit
tqdm
matplotlib
scikit-image
